Combine all remote watch functionality into remote_watch.go
