+++ pkg/controller/migcluster/remote_watch.go
@@ -18,6 +18,7 @@ package migcluster
 
 import (
 	"os"
+	"sync"
 
 	"github.com/fusor/mig-controller/pkg/controller/remotewatcher"
 	"k8s.io/apimachinery/pkg/types"
@@ -30,6 +31,9 @@ import (
 	migrationv1alpha1 "github.com/fusor/mig-controller/pkg/apis/migration/v1alpha1"
 	velerov1 "github.com/heptio/velero/pkg/apis/velero/v1"
 )
+
+var rwmInstance *RemoteWatchMap
+var createRemoteWatchMapOnce sync.Once
 
 // RemoteManagerConfig specifies config options for setting up a RemoteWatch Manager
 type RemoteManagerConfig struct {
@@ -99,4 +103,51 @@ func StartRemoteWatch(r *ReconcileMigCluster, config RemoteManagerConfig) error 
 	rwm.SetRWC(config.ParentNsName, rwc)
 
 	return nil
+}
+
+// RemoteWatchCluster tracks Remote Managers and Event Forward Channels
+type RemoteWatchCluster struct {
+	ForwardChannel chan event.GenericEvent
+	RemoteManager  manager.Manager
+	//  TODO - setup stop channel for manager so that manager will stop when we close the event channel from parent
+}
+
+// RemoteWatchMap provides a map between MigCluster nsNames and RemoteWatchClusters
+type RemoteWatchMap struct {
+	sync.RWMutex
+	nsNameToRemoteCluster map[types.NamespacedName]*RemoteWatchCluster
+}
+
+// GetRemoteWatchMap returns the shared RemoteWatchMap instance
+func GetRemoteWatchMap() *RemoteWatchMap {
+	createRemoteWatchMapOnce.Do(func() {
+		rwmInstance = &RemoteWatchMap{}
+		rwmInstance.nsNameToRemoteCluster = make(map[types.NamespacedName]*RemoteWatchCluster)
+	})
+	return rwmInstance
+}
+
+// GetRWC gets the RemoteWatchCluster associated with a MigCluster resource nsName, return nil if not found
+func (r *RemoteWatchMap) GetRWC(key types.NamespacedName) *RemoteWatchCluster {
+	r.RLock()
+	defer r.RUnlock()
+	rwc, ok := rwmInstance.nsNameToRemoteCluster[key]
+	if !ok {
+		return nil
+	}
+	return rwc
+}
+
+// SetRWC sets the RemoteWatchCluster associated with a MigCluster resource nsName
+func (r *RemoteWatchMap) SetRWC(key types.NamespacedName, value *RemoteWatchCluster) {
+	r.Lock()
+	defer r.Unlock()
+	rwmInstance.nsNameToRemoteCluster[key] = value
+}
+
+// DeleteRWC deletes the RemoteWatchCluster associated with a MigCluster resource nsName
+func (r *RemoteWatchMap) DeleteRWC(key types.NamespacedName) {
+	r.Lock()
+	defer r.Unlock()
+	delete(rwmInstance.nsNameToRemoteCluster, key)
 }
